refactor(conference): use raw events and map later
We'll need to pass a few things down through signaling instead, and this will remove the need to have additional `ConferenceEvent`s
