+++ sdk-conference-infinity/src/main/kotlin/com/pexip/sdk/conference/infinity/internal/ConferenceEvent.kt
@@ -16,28 +16,19 @@
 package com.pexip.sdk.conference.infinity.internal
 
 import com.pexip.sdk.api.Event
-import com.pexip.sdk.api.coroutines.asFlow
 import com.pexip.sdk.api.infinity.DisconnectEvent
-import com.pexip.sdk.api.infinity.InfinityService
 import com.pexip.sdk.api.infinity.MessageReceivedEvent
 import com.pexip.sdk.api.infinity.PresentationStartEvent
 import com.pexip.sdk.api.infinity.PresentationStopEvent
 import com.pexip.sdk.api.infinity.ReferEvent
-import com.pexip.sdk.api.infinity.TokenStore
 import com.pexip.sdk.conference.DisconnectConferenceEvent
 import com.pexip.sdk.conference.FailureConferenceEvent
 import com.pexip.sdk.conference.MessageReceivedConferenceEvent
 import com.pexip.sdk.conference.PresentationStartConferenceEvent
 import com.pexip.sdk.conference.PresentationStopConferenceEvent
 import com.pexip.sdk.conference.ReferConferenceEvent
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.flatMapLatest
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.mapNotNull
-import kotlinx.coroutines.flow.retryWhen
-import kotlin.time.Duration.Companion.seconds
 
+@Suppress("ktlint:standard:function-naming")
 internal inline fun ConferenceEvent(
     event: Event,
     at: () -> Long = System::currentTimeMillis,
@@ -67,17 +58,6 @@ internal inline fun ConferenceEvent(
     else -> null
 }
 
+@Suppress("ktlint:standard:function-naming")
 internal inline fun ConferenceEvent(t: Throwable, at: () -> Long = System::currentTimeMillis) =
     FailureConferenceEvent(at(), t)
-
-@OptIn(ExperimentalCoroutinesApi::class)
-internal fun InfinityService.ConferenceStep.conferenceEvent(
-    store: TokenStore,
-    at: () -> Long = System::currentTimeMillis,
-) = flow { emit(store.get()) }
-    .flatMapLatest { events(it).asFlow() }
-    .mapNotNull { ConferenceEvent(it, at) }
-    .retryWhen { _, attempt ->
-        delay(attempt.seconds.coerceAtMost(5.seconds))
-        true
-    }
