+++ sdk-conference-infinity/src/main/kotlin/com/pexip/sdk/conference/infinity/internal/Events.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2023 Pexip AS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.pexip.sdk.conference.infinity.internal
+
+import com.pexip.sdk.api.coroutines.asFlow
+import com.pexip.sdk.api.infinity.InfinityService
+import com.pexip.sdk.api.infinity.TokenStore
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.retryWhen
+import kotlin.time.Duration.Companion.seconds
+
+@OptIn(ExperimentalCoroutinesApi::class)
+internal fun InfinityService.ConferenceStep.events(store: TokenStore) = flow { emit(store.get()) }
+    .flatMapLatest { events(it).asFlow() }
+    .retryWhen { _, attempt ->
+        delay(attempt.seconds.coerceAtMost(5.seconds))
+        true
+    }
