+++ sdk-conference-infinity/src/test/kotlin/com/pexip/sdk/conference/infinity/internal/ConferenceEventTest.kt
@@ -15,40 +15,20 @@
  */
 package com.pexip.sdk.conference.infinity.internal
 
-import app.cash.turbine.test
-import assertk.Assert
-import assertk.all
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import assertk.assertions.isInstanceOf
-import assertk.assertions.prop
 import com.pexip.sdk.api.Event
-import com.pexip.sdk.api.EventSource
-import com.pexip.sdk.api.EventSourceFactory
-import com.pexip.sdk.api.EventSourceListener
-import com.pexip.sdk.api.infinity.ByeEvent
 import com.pexip.sdk.api.infinity.DisconnectEvent
 import com.pexip.sdk.api.infinity.MessageReceivedEvent
 import com.pexip.sdk.api.infinity.PresentationStartEvent
 import com.pexip.sdk.api.infinity.PresentationStopEvent
 import com.pexip.sdk.api.infinity.ReferEvent
-import com.pexip.sdk.api.infinity.Token
 import com.pexip.sdk.api.infinity.TokenStore
-import com.pexip.sdk.conference.ConferenceEvent
 import com.pexip.sdk.conference.DisconnectConferenceEvent
 import com.pexip.sdk.conference.FailureConferenceEvent
 import com.pexip.sdk.conference.MessageReceivedConferenceEvent
 import com.pexip.sdk.conference.PresentationStartConferenceEvent
 import com.pexip.sdk.conference.PresentationStopConferenceEvent
 import com.pexip.sdk.conference.ReferConferenceEvent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.test.TestScope
-import kotlinx.coroutines.test.runTest
 import java.util.UUID
 import kotlin.properties.Delegates
 import kotlin.random.Random
@@ -68,78 +48,6 @@ class ConferenceEventTest {
         at = Random.nextLong(Long.MAX_VALUE)
         event = MutableSharedFlow()
         store = TokenStore.create(Random.nextToken())
-    }
-
-    @Test
-    fun `maps Event to ConferenceEvent`() = runTest {
-        val step = testConferenceStep()
-        step.conferenceEvent(store) { at }.test {
-            event.awaitSubscribers()
-            val presentationStartEvent = PresentationStartEvent(
-                presenterId = UUID.randomUUID(),
-                presenterName = Random.nextString(8),
-            )
-            event.emit(Result.success(presentationStartEvent))
-            assertThat(awaitItem(), "presentationStartEvent")
-                .isInstanceOf<PresentationStartConferenceEvent>()
-                .all {
-                    at(at)
-                    prop(PresentationStartConferenceEvent::presenterId)
-                        .isEqualTo(presentationStartEvent.presenterId)
-                    prop(PresentationStartConferenceEvent::presenterName)
-                        .isEqualTo(presentationStartEvent.presenterName)
-                }
-            val presentationStopEvent = PresentationStopEvent
-            event.emit(Result.success(presentationStopEvent))
-            assertThat(awaitItem(), "presentationStopEvent")
-                .isInstanceOf<PresentationStopConferenceEvent>()
-                .at(at)
-            val referEvent = ReferEvent(
-                conferenceAlias = Random.nextString(8),
-                token = Random.nextString(8),
-            )
-            event.emit(Result.success(referEvent))
-            assertThat(awaitItem(), "referEvent")
-                .isInstanceOf<ReferConferenceEvent>()
-                .all {
-                    at(at)
-                    prop(ReferConferenceEvent::conferenceAlias).isEqualTo(referEvent.conferenceAlias)
-                    prop(ReferConferenceEvent::token).isEqualTo(referEvent.token)
-                }
-            val disconnectEvent = DisconnectEvent(Random.nextString(8))
-            event.emit(Result.success(disconnectEvent))
-            assertThat(awaitItem(), "disconnectEvent")
-                .isInstanceOf<DisconnectConferenceEvent>()
-                .all {
-                    at(at)
-                    prop(DisconnectConferenceEvent::reason).isEqualTo(disconnectEvent.reason)
-                }
-            val byeEvent = ByeEvent
-            event.emit(Result.success(byeEvent))
-            expectNoEvents()
-            cancelAndIgnoreRemainingEvents()
-        }
-    }
-
-    @Test
-    fun `failure restarts the flow`() = runTest {
-        val step = testConferenceStep()
-        step.conferenceEvent(store) { at }.test {
-            event.awaitSubscribers()
-            val presentationStopEvent = PresentationStopEvent
-            event.emit(Result.success(presentationStopEvent))
-            assertThat(awaitItem(), "presentationStopEvent")
-                .isInstanceOf<PresentationStopConferenceEvent>()
-                .at(at)
-            event.emit(Result.failure(Throwable()))
-            event.awaitSubscribers()
-            event.emit(Result.success(presentationStopEvent))
-            assertThat(awaitItem(), "presentationStopEvent")
-                .isInstanceOf<PresentationStopConferenceEvent>()
-                .at(at)
-            expectNoEvents()
-            cancelAndIgnoreRemainingEvents()
-        }
     }
 
     @Test
@@ -200,45 +108,4 @@ class ConferenceEventTest {
     }
 
     private object TestEvent : Event
-
-    private suspend fun <T> MutableSharedFlow<T>.awaitSubscribers() {
-        subscriptionCount.first { it > 0 }
-    }
-
-    private fun TestScope.testConferenceStep() = object : TestConferenceStep() {
-
-        override fun events(token: Token): EventSourceFactory {
-            assertThat(token, "token").isEqualTo(store.get())
-            return TestEventSourceFactory(backgroundScope, event)
-        }
-    }
-
-    private class TestEventSourceFactory(
-        private val scope: CoroutineScope,
-        private val event: Flow<Result<Event>>,
-    ) : EventSourceFactory {
-
-        override fun create(listener: EventSourceListener): EventSource =
-            TestEventSource(scope, event, listener)
-    }
-
-    private class TestEventSource(
-        scope: CoroutineScope,
-        event: Flow<Result<Event>>,
-        listener: EventSourceListener,
-    ) : EventSource {
-
-        private val job = event
-            .onEach {
-                it.fold(
-                    onSuccess = { event -> listener.onEvent(this, event) },
-                    onFailure = { t -> listener.onClosed(this, t) },
-                )
-            }
-            .launchIn(scope)
-
-        override fun cancel(): Unit = job.cancel()
-    }
-
-    private fun Assert<ConferenceEvent>.at(at: Long) = prop(ConferenceEvent::at).isEqualTo(at)
 }
